Split pkg_elf into pkg_abi (generic part) and pkg_elf (ELF specific part)
