+++ libpkg/Makefile.autosetup
@@ -32,6 +32,7 @@ SRCS=	backup_lib.c \
 	pkg_jobs_conflicts.c pkg_ports.c \
 	pkg_solve.c \
 	pkgdb.c \
+	pkg_abi.c \
 	pkg_elf.c \
 	binfmt_macho.c \
 	ssh.c elfhints.c \
