+++ libpkg/private/pkg.h
@@ -640,8 +640,7 @@ struct os_info {
 	char *arch;
 };
 
-int pkg_get_myarch(char *pkgarch, size_t sz, struct os_info *);
-int pkg_get_myarch_legacy(char *pkgarch, size_t sz);
+int pkg_get_myarch_with_legacy(char *pkgarch, char* legacy, size_t sz, struct os_info *);
 
 /**
  * Remove and unregister the package.
